Moving tests to right place
